% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{getty_catalogs}
\alias{getty_catalogs}
\title{Getty Catalogue Digitization Progress}
\format{A data frame with 8393 observations of 4 variables:
\describe{
 \item{sale_begin}{POSIXct. The start date for the sale described by the catalog.}
 \item{sale_no}{Character. The sale catalog id number.}
 \item{is_input}{Logical. Has the catalog been input into the database yet?}
 \item{year}{Integer. Year derived from \code{sale_begin}}
}}
\source{
Corresepondence with Christian Huemer, August 2015.
}
\usage{
getty_catalogs
}
\description{
A table describing the Getty's progress in digitizing their sales catalogs.
}
\keyword{datasets}

