% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/londonauctions-package.r
\docType{package}
\name{londonauctions}
\alias{londonauctions}
\alias{londonauctions-package}
\title{londonauctions.}
\description{
Included data: \link{sales}, \link{ra_dates}, \link{getty_catalogs}, and \link{parliament_dates}.
}

