% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parliament_dates}
\alias{parliament_dates}
\title{Parliament Dates}
\format{A data frame with 128 observations of 6 variables:
\describe{
 \item{par_start}{POSIXct. The date of the parliament start}
 \item{par_end}{POSIXct. The date of the parliament end}
 \item{par_start_year}{Integer. Year parliament began.}
 \item{par_end_year}{Integer. Year parliament ended.}
 \item{par_start_day}{Integer. Day of the year the parliament began.}
 \item{par_end_day}{Integer. Day of the year the parliament ended.}
}}
\usage{
parliament_dates
}
\description{
A table of the starting and ending dates for Parliament, 1715-1831.
}
\keyword{datasets}

