% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peaks}
\alias{peaks}
\title{Determine CV trends for a given set of sales}
\usage{
peaks(df)
}
\arguments{
\item{df}{Filtered data frame of sales}

\item{n_days}{Number of top days to measure}
}
\value{
A data frame
}
\description{
Determine CV trends for a given set of sales
}

