% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_dates}
\alias{ra_dates}
\title{Royal Academy Exhibition Dates}
\format{A data frame with 83 observations of 4 variables:
\describe{
 \item{exhib_start}{POSIXct. The date of the exhibition start}
 \item{exhib_end}{POSIXct. The date of the exhibition end}
 \item{exhib_start_day}{Integer. Day of the year the exhibition began.}
 \item{exhib_end_day}{Integer. Day of the year the exhibition ended.}
 \item{exhib_year}{Integer. Year of the exhibition.}
}}
\usage{
ra_dates
}
\description{
A table of the Royal Academy exhibiton start and end days, as compiled by RA archivist Mark Pomeroy
}
\keyword{datasets}

