% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales}
\alias{sales}
\title{London Paintings Auction Records}
\format{A data frame with 12 variables:
\describe{
 \item{cat_no}{Character. Auction catalogue number}
 \item{lot_no}{Character. Lot number}
 \item{auction_house}{Character. The auction house associated with the sale
 record}
 \item{date}{POSIXct. The date of the lot sale. Where a specific lot sale
 date is specified, that date is used. Where only a range of dates is
 specified, the earliest date is used.}
 \item{transaction_type}{Character. The transaction type.}
 \item{transaction_amt}{Numeric. Transaciton amount.}
 \item{year}{Integer.}
 \item{month}{Integer. Month of the year.}
 \item{week}{Integer. Week of the year.}
 \item{yday}{Integer. Day of the year.}
 \item{mday}{Integer. Day of the month.}
 \item{wday}{Integer. Day of the week.}
 \item{transaction}{Character. Verbatim text of the transaction from the Getty catalogs.}
 \item{artists}{Character. Verbatim text of artist names from the Getty catalogs. Multiple names are delimited with a semicolon.}
 \item{title}{Character. Verbatim title of the artwork from the Getty catalogs.}
 \item{sellers}{Character. Verbatim text of seller names from the Getty catalogs. Multiple names are delimited with a semicolon.}
 \item{buyers}{Character. Verbatim text of buyer names from the Getty catalogs. Multiple names are delimited with a semicolon.}
}}
\usage{
sales
}
\description{
A cleaned database of paintings records derived from the Getty Provenance Index.
}
\keyword{datasets}

